Create bitwise.c
You will be given an integer n, and the threshold, k. For each number i from 1 through n that are maximum value of the logical and, or and xor when compared against all the integers through n that are greater than i. consider a value only if the comparison returns a result less than k. print the results of the and ,or and exclusive or comparisons on the separate lines, in that order.

Function Description

Complete the calculate_the_maximum function in the editor below.

calculate_the_maximum has the following parameters:

int n: the highest number to consider
int k: the result of a comparison must be lower than this number to be considered.