bitwise.c
You will be given an integer n, and a threshold k. For each number i from 1 through n, find the maximum value of the logical and, or and xor when compared against all the integer through n that are greater than i. Consider a value only if the comparison returns a result less than k. Print the results of the bitwise operators in the and, or and xor order.

Function Description

Complete the calculate_the_maximum function in the editor below.

calculate_the_maximum has the following parameters:

int n: the highest number to consider
int k: the result of a comparison must be lower than this number to be considered